requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
langchain==0.0.335
langchain-openai==0.0.2
langchain-google-genai==0.0.5
langchain-community==0.0.8
langchain-groq==0.0.1
groq==0.4.1
langgraph==0.0.26
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
requests==2.31.0
pytz==2023.3
datetime
typing-extensions==4.8.0